% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_get.R
\name{api_get}
\alias{api_get}
\title{Get interactions from the COVIDminer database.}
\usage{
api_get(
  query,
  filters = "all",
  level = 1,
  unmapped = TRUE,
  meta = FALSE,
  sort = FALSE,
  timeout = 100,
  show.url = FALSE
)
}
\arguments{
\item{query}{A character vector of text descriptors of database IDs. See
Details.}

\item{filters}{A character vector of filters. Default is \code{all}, which
will not apply any filters. A good option for general use is
\code{default}, which will only retrieve SARS-COV-2 primary literature.}

\item{level}{An integer describing the network level that is to be returned. 
Either 0 (only the entities specified in \code{query}) or 1 (the query entities and their first neighbours; the default).}

\item{unmapped}{Logical. Should unmapped entities be used?}

\item{meta}{Logical. If \code{TRUE}, duplicate interactions will be merged
into a single directed edge. The results column \code{edge_weight} contains
the number of edges that were merged. Default is to show each interaction
separately (so that literature source information is retained).}

\item{sort}{Logical. Should the result be sorted. Currently not implemented.}

\item{timeout}{An integer specifying the timeout in seconds for the http
query.}

\item{show.url}{Logical. Should the URL be displayed in the console (useful
for debugging).}
}
\value{
A nested \code{list} object with top level elements; query (details
  of the query as interpreted by the server), results (a \code{data.frame}
  with results for each query element).
}
\description{
This function takes a vector of COVIDminer database IDs and allows networks
containing these entities to be retrieved.
}
\examples{

search.response <- COVIDminerAPI::api_search(query = c("bdnf", "autophagy"))
best.matches <- search.response$results$`best-match`

get.response <- COVIDminerAPI::api_get(query = best.matches, filters = "default")
edge.list <- get.response$results[, c("source", "relation", "target")]

}
