% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_search.R
\name{api_search}
\alias{api_search}
\title{Search the COVIDminer database for matching identifiers.}
\usage{
api_search(query, sort = FALSE, timeout = 100)
}
\arguments{
\item{query}{A character vector of text descriptors of database IDs. See
Details.}

\item{sort}{Logical. Should the result be sorted. Currently not implemented.}

\item{timeout}{An integer specifying the timeout in seconds for the http
query.}
}
\value{
A nested \code{list} object with top level elements; query (details
  of the query as interpreted by the server), results (a \code{data.frame}
  with results for each query element), matches; all matching database
  entities for each query element) and hits; the COVIDminer IDs where a
  perfect match was found).
}
\description{
This function takes free text values and searches the COVIDminer database for
entry IDs, names or aliases with partial matches.
}
\examples{

search.response <- COVIDminerAPI::api_search(query = c("bdnf", "autophagy"))
best.matches <- search.response$results$`best-match`

get.response <- COVIDminerAPI::api_get(query = best.matches, filters = "default")
edge.list <- get.response$results[, c("source", "relation", "target")]

}
